Force paths that are passed to tests to be in canonical form so that the backup tests will pass. Fix a potential NPE in the backup test.
svn path=/trunk/; revision=46173


Former-commit-id: 6b5c8d5f48d5a5daca43555ec7d26a89e9f67bb0

+++ v2/appserv-tests/devtests/admin/cli/src/admin/BackupCommandsTest.java
@@ -161,8 +161,11 @@ public class BackupCommandsTest extends AdminBaseDevTest {
 
     private void cleanupBackupDir() {
         File path = new File(BACKUP_DIR, DOMAIN1);
-        for(File f : path.listFiles()) {
-            f.delete();
+        File[] paths = path.listFiles();
+        if (paths != null) {
+            for(File f : paths) {
+                f.delete();
+            }
         }
         path.delete();
     }
